+++ src/interface/index.ts
@@ -1,1 +1,3 @@
 export { InterfaceModule } from './interface.module';
+
+export { type ScheduledJob, ScheduledJobsToken, SchedulerModule } from './scheduler';